requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
Brotli==1.1.0
discord.py==2.3.2
frozenlist==1.4.1
idna==3.7
inflate64==1.0.0
multidict==6.0.5
multivolumefile==0.2.3
psutil==5.9.8
py7zr==0.21.0
pybcj==1.0.2
pycryptodomex==3.20.0
pyppmd==1.1.0
pyzstd==0.15.10
texttable==1.7.0
yarl==1.9.4
